Add advanced filtering, quick actions, and improved image fallback
Introduces ActiveFilters and FilterSidebar components for advanced filtering and sorting, and a QuickActionsMenu for marking items as watched, wishlist, or adding to lists. Enhances ContentCard with quick actions and improved badges, updates ImageWithFallback to use a local fallback image, and improves SeriesSeasons with release checks and unreleased notifications. Adds new API route for now playing movies, utility for release date validation, and various UI/UX improvements across components.
And more

+++ watch-hive/src/app/utils/dateFormatter.js
@@ -1,5 +1,5 @@
 /**
- * Format date string to readable format (e.g., "Nov 11, 2025")
+ * Format date string to readable format (e.g., "November 11, 2025")
  */
 export function formatDate(dateString) {
     if (!dateString) return 'N/A';
@@ -12,7 +12,7 @@ export function formatDate(dateString) {
         
         return date.toLocaleDateString('en-US', {
             year: 'numeric',
-            month: 'short',
+            month: 'long',
             day: 'numeric'
         });
     } catch (error) {
